+++ Deflib/MLdata.cs
@@ -9,21 +9,21 @@ namespace Deflib
     //All functions used in ML written in C# for when testing data
     public class Generate_data
     {
-        public static double[] csvfile(string file)
+        public static float[] csvfile(string file)
         {
             var A = File.ReadAllLines(file)
                 .Where(x => !string.IsNullOrWhiteSpace(x))
-                .Select(x => double.Parse(x.Trim(), CultureInfo.InvariantCulture))
+                .Select(x => float.Parse(x.Trim(), CultureInfo.InvariantCulture))
                 .ToArray();
 
             return A;
         }
 
-        public static double[,] csvfile(string file, int x, int y)
+        public static float[,] csvfile(string file, int x, int y)
         {
             var A = File.ReadAllLines(file)
                 .Where(x => !string.IsNullOrWhiteSpace(x))
-                .Select(x => double.Parse(x.Trim(), CultureInfo.InvariantCulture))
+                .Select(x => float.Parse(x.Trim(), CultureInfo.InvariantCulture))
                 .ToArray();
 
             var Reshape = Functions.reshape(A, x, y);
@@ -31,11 +31,11 @@ namespace Deflib
             return Reshape;
         }
 
-        public static double[,,] csvfile(string file, int x, int y, int z)
+        public static float[,,] csvfile(string file, int x, int y, int z)
         {
             var A = File.ReadAllLines(file)
                 .Where(x => !string.IsNullOrWhiteSpace(x))
-                .Select(x => double.Parse(x.Trim(), CultureInfo.InvariantCulture))
+                .Select(x => float.Parse(x.Trim(), CultureInfo.InvariantCulture))
                 .ToArray();
 
             var Reshape = Functions.reshape(A, x, y, z);
@@ -43,7 +43,7 @@ namespace Deflib
             return Reshape;
         }
 
-        public static double[,] matmul_mat(double[,] x, double[,,] W0)
+        public static float[,] matmul_mat(float[,] x, float[,,] W0)
         {
             var hh = Functions.matmul(x,
                 Functions.transpose(
@@ -57,7 +57,7 @@ namespace Deflib
             return hh;
         }
 
-        public static Tuple<double[,,], double[,,]> generate_hz(double[,] hh, double[,] x)
+        public static Tuple<float[,,], float[,,]> generate_hz(float[,] hh, float[,] x)
         {
             var batchsize = x.GetLength(0);
             var h = Functions.reshape(hh, batchsize, (int)Parameters.num_networks, (int)Parameters.hidden_size);
@@ -65,7 +65,7 @@ namespace Deflib
             int b = h.GetLength(1);
             int c = h.GetLength(2);
 
-            double[,,] hz = new double[a,b,c];
+            float[,,] hz = new float[a,b,c];
             for (int i = 0; i < a; i++)
             {
                 for (int j = 0; j < b; j++)
@@ -84,7 +84,7 @@ namespace Deflib
                 }
             }
 
-            double[,,] hr = new double[a,b,c];
+            float[,,] hr = new float[a,b,c];
             for (int i = 0; i < a; i++)
             {
                 for (int j = 0; j < b; j++)
@@ -103,16 +103,16 @@ namespace Deflib
                 }
             }
 
-            return new Tuple<double[,,], double[,,]>(hz, hr);
+            return new Tuple<float[,,], float[,,]>(hz, hr);
         }
 
-        public static Tuple<double[,,], double[,,]> z_r(double[,,] hz, double[,,] hr )
+        public static Tuple<float[,,], float[,,]> z_r(float[,,] hz, float[,,] hr )
         {
             int a = hz.GetLength(0);
             int b = hz.GetLength(1);
             int c = hz.GetLength(2);
 
-            double[,,] wz = new double[a,b,c];
+            float[,,] wz = new float[a,b,c];
             for (int i = 0; i < a; i++)
             {
                 for (int j = 0; j < b; j++)
@@ -124,7 +124,7 @@ namespace Deflib
                 }
             }
 
-            double[,,] z = new double[a,b,c];
+            float[,,] z = new float[a,b,c];
             Functions.ZeroMatrix(z);
             for (int i = 0; i < a; i++)
             {
@@ -141,7 +141,7 @@ namespace Deflib
             int f = hr.GetLength(1);
             int g = hr.GetLength(2);
 
-            double[,,] wr = new double[e,f,g];
+            float[,,] wr = new float[e,f,g];
             for (int i = 0; i < e; i++)
             {
                 for (int j = 0; j < f; j++)
@@ -153,7 +153,7 @@ namespace Deflib
                 }
             }
 
-            double[,,] r = new double[e,f,g];
+            float[,,] r = new float[e,f,g];
             Functions.ZeroMatrix(r);
             for (int i = 0; i < e; i++)
             {
@@ -166,15 +166,15 @@ namespace Deflib
                 }
             }
 
-            return new Tuple<double[,,], double[,,]>(z, r);
+            return new Tuple<float[,,], float[,,]>(z, r);
         }
 
-        public static Tuple<double[,], double[,]> gen_SLA(double[,,] z, double[,,] r)
+        public static Tuple<float[,], float[,]> gen_SLA(float[,,] z, float[,,] r)
         {
             int a = z.GetLength(0);
             int b = z.GetLength(1);
             int c = z.GetLength(2);
-            double[,] sum_z_lastaxis = new double[a,b];
+            float[,] sum_z_lastaxis = new float[a,b];
             Functions.ZeroMatrix(sum_z_lastaxis);
             for (int i = 0; i < a; i++)
             {
@@ -190,7 +190,7 @@ namespace Deflib
             int e = r.GetLength(0);
             int f = r.GetLength(1);
             int g = r.GetLength(2);
-            double[,] sum_r_lastaxis = new double[e,f];
+            float[,] sum_r_lastaxis = new float[e,f];
             Functions.ZeroMatrix(sum_r_lastaxis);
             for (int i = 0; i < e; i++)
             {
@@ -203,15 +203,15 @@ namespace Deflib
                 }
             }
 
-            return new Tuple<double[,], double[,]>(sum_z_lastaxis, sum_r_lastaxis);
+            return new Tuple<float[,], float[,]>(sum_z_lastaxis, sum_r_lastaxis);
         }
 
-        public static double[,] zz (double[,] z)
+        public static float[,] zz (float[,] z)
         {
             int a = z.GetLength(0);
             int b = z.GetLength(1);
 
-            double[,] zz_scale = new double[a,b];
+            float[,] zz_scale = new float[a,b];
             for (int i = 0; i < a; i++)
             {
                 for (int j = 0; j < b; j++)
@@ -225,12 +225,12 @@ namespace Deflib
             return zz;
         }
 
-        public static double[,] sig (double[,] zz)
+        public static float[,] sig (float[,] zz)
         {
             return Functions.sigmoid2d(zz);
         }
 
-        public static double[,] mulmin (double[,] zz_sig)
+        public static float[,] mulmin (float[,] zz_sig)
         {
             int a = zz_sig.GetLength(0);
             int b = zz_sig.GetLength(1);
@@ -247,25 +247,25 @@ namespace Deflib
             return zz_sig;
         }
 
-        public static double[,] soft(double[,] r)
+        public static float[,] soft(float[,] r)
         {
             return Functions.softplus(r);
         }
 
-        public static double[,] rz(double[,] soft,double[,] mulmin)
+        public static float[,] rz(float[,] soft,float[,] mulmin)
         {
             return Functions.multiply(soft,mulmin);
         }
 
-        public static double[,] clamp(double[,] rz)
+        public static float[,] clamp(float[,] rz)
         {
             return Functions.clamp(rz, - (int)Parameters.max_predict, (int)Parameters.max_predict);
         }
 
-        public static double[] ensemble_predictions(double[,] clamp)
+        public static float[] ensemble_predictions(float[,] clamp)
         {
             var y_mean = Functions.mean(clamp);
-            double[] mean = new double[1] {y_mean};
+            float[] mean = new float[1] {y_mean};
 
             return mean;
         }
