+++ HzHr/simulation.cs
@@ -6,7 +6,7 @@ namespace HzHr
 
     public class Generate_matmul
     {
-        public static double[,] Matmul_generate()
+        public static float[,] Matmul_generate()
         {
             return Deflib.Generate_data.matmul_mat(Deflib.dataMatrix.x, Deflib.dataMatrix.W0);
         }
